Combine 1.2.5 tests
