--- src/test/kotlin/xyz/wagyourtail/unimined/test/integration/ForgeModloader1_2_5Test.kt
@@ -1,30 +1,0 @@
-package xyz.wagyourtail.unimined.test.integration
-
-import org.gradle.testkit.runner.TaskOutcome
-import org.gradle.testkit.runner.UnexpectedBuildFailure
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.MethodSource
-import xyz.wagyourtail.unimined.util.runTestProject
-
-class ForgeModloader1_2_5Test {
-    @ParameterizedTest
-    @MethodSource("xyz.wagyourtail.unimined.util.IntegrationTestUtils#versions")
-    fun test_forge_modloader_1_2_5(gradleVersion: String) {
-        try {
-            val result = runTestProject("1.2.5-Forge-Modloader", gradleVersion)
-
-            try {
-                result.task(":build")?.outcome?.let {
-                    if (it != TaskOutcome.SUCCESS) throw Exception("build failed")
-                } ?: throw Exception("build failed")
-            } catch (e: Exception) {
-                println(result.output)
-                throw Exception(e)
-            }
-        } catch (e: UnexpectedBuildFailure) {
-            println(e)
-            throw Exception("build failed", e)
-        }
-    }
-}
