--- testing/1.2.5-Fabric-Quilt/build.gradle
@@ -1,110 +1,0 @@
-plugins {
-    id 'java'
-    id 'xyz.wagyourtail.unimined' // version '1.0.0'
-}
-
-group 'com.example'
-version '1.0-SNAPSHOT'
-
-base {
-    archivesName = 'UniminedExampleMod'
-}
-
-// this is just here so we can test the outputs easier and clean between tests
-unimined.useGlobalCache = false
-
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-sourceSets {
-    fabric {
-        compileClasspath += sourceSets.main.output
-        runtimeClasspath += sourceSets.main.output
-    }
-    quilt {
-        compileClasspath += sourceSets.main.output + sourceSets.fabric.output
-        runtimeClasspath += sourceSets.main.output + sourceSets.fabric.output
-    }
-}
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-}
-
-unimined.minecraft() {
-    version "1.2.5"
-    side "client"
-
-    mappings {
-        calamus()
-        feather(20)
-
-        devFallbackNamespace "intermediary"
-    }
-
-    minecraftRemapper.config {
-        ignoreConflicts(true)
-    }
-
-    defaultRemapJar = false
-
-}
-
-unimined.minecraft(sourceSets.fabric) {
-    version "1.2.5"
-    side "client"
-
-    mappings {
-        calamus()
-        feather(20)
-    }
-
-    minecraftRemapper.config {
-        ignoreConflicts(true)
-    }
-
-    fabric {
-        loader '0.14.24'
-    }
-
-    defaultRemapJar = false
-}
-
-unimined.minecraft(sourceSets.quilt) {
-    version "1.2.5"
-    side "client"
-
-    mappings {
-        calamus()
-        feather(20)
-    }
-
-    minecraftRemapper.config {
-        ignoreConflicts(true)
-    }
-
-    quilt {
-        loader '0.21.2'
-    }
-
-    defaultRemapJar = false
-}
-
-repositories {
-}
-
-dependencies {
-}
-
-jar {
-    enabled = false
-}
-
-processResources {
-    inputs.property "version", project.version
-
-    filesMatching("fabric.mod.json") {
-        expand "version": project.version
-    }
-}
