--- testing/1.2.5-Fabric-Quilt/gradle.properties
@@ -1,1 +1,0 @@
-org.gradle.jvmargs=-Xmx2G
