--- testing/1.2.5-Fabric-Quilt/settings.gradle
@@ -1,30 +1,0 @@
-pluginManagement {
-    repositories {
-        mavenCentral()
-        maven {
-            url = "https://maven.neoforged.net/releases"
-        }
-        maven {
-            url = "https://maven.minecraftforge.net/"
-        }
-        maven {
-            url = "https://maven.fabricmc.net/"
-        }
-        maven {
-            url = "https://maven.wagyourtail.xyz/releases"
-        }
-        maven {
-            url = "https://maven.wagyourtail.xyz/snapshots"
-        }
-        gradlePluginPortal() {
-            content {
-                excludeGroup("org.apache.logging.log4j")
-            }
-        }
-    }
-}
-
-// so we can use the unimined directly provided by the super project
-includeBuild('../../')
-
-rootProject.name = '1.2.5-Fabric-Quilt'
