--- testing/1.2.5-Fabric-Quilt/src/fabric/java/com/example/fabric/ExampleModFabric.java
@@ -1,14 +1,0 @@
-package com.example.fabric;
-
-import net.fabricmc.api.ModInitializer;
-
-import java.util.logging.Logger;
-
-public class ExampleModFabric implements ModInitializer {
-	public static Logger LOGGER = Logger.getLogger("ExampleMod");
-
-	@Override
-	public void onInitialize() {
-		LOGGER.info("Hello from Fabric!");
-	}
-}
