--- testing/1.2.5-Fabric-Quilt/src/fabric/resources/fabric.mod.json
@@ -1,29 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "modid",
-  "version": "${version}",
-  "name": "Example Mod",
-  "description": "This is an example description! Tell everyone what your mod is about!",
-  "authors": [
-    "Me!"
-  ],
-  "contact": {
-    "homepage": "https://fabricmc.net/",
-    "sources": "https://github.com/FabricMC/fabric-example-mod"
-  },
-  "license": "CC0-1.0",
-  "environment": "*",
-  "entrypoints": {
-    "main": [
-      "com.example.fabric.ExampleModFabric"
-    ]
-  },
-  "mixins": [
-    "modid.mixins.json"
-  ],
-  "depends": {
-    "fabricloader": ">=0.14.19",
-    "minecraft": "1.2.5",
-    "java": ">=8"
-  }
-}
