--- testing/1.2.5-Forge-Modloader/gradle.properties
@@ -1,1 +1,0 @@
-org.gradle.jvmargs=-Xmx2G
