--- testing/1.2.5-Forge-Modloader/src/modloader/resources/exampleModId.transform
@@ -1,1 +1,0 @@
-com.example.transform.GuiMainMenuTransform
