+++ testing/minecraft/1.2.5/gradle.properties
@@ -1,0 +1,7 @@
+org.gradle.jvmargs=-Xmx2G
+
+minecraft_version=1.2.5
+mcp_version=6.2
+forge_version=3.4.9.171
+fabric_version=0.16.0
+quilt_version=0.26.3
