+++ testing/minecraft/1.2.5/src/fabric/java/com/example/example/mod/FabricExampleMod.java
@@ -1,0 +1,12 @@
+package com.example.example.mod;
+
+import net.fabricmc.api.ModInitializer;
+
+import java.util.logging.Logger;
+
+public class FabricExampleMod implements ModInitializer {
+	@Override
+	public void onInitialize() {
+		ExampleMod.LOGGER.info("Hello Fabric world!");
+	}
+}
