+++ testing/minecraft/1.2.5/src/main/java/com/example/example/mod/ExampleMod.java
@@ -1,0 +1,7 @@
+package com.example.example.mod;
+
+import java.util.logging.*;
+
+public class ExampleMod {
+	public static Logger LOGGER = Logger.getLogger("ExampleMod");
+}
