+++ testing/minecraft/1.2.5/src/modloader/java/com/example/example/mod/mod_ExampleMod.java
@@ -1,0 +1,31 @@
+package com.example.example.mod;
+
+import net.minecraft.src.BaseMod;
+
+import java.util.*;
+
+public class mod_ExampleMod extends BaseMod {
+    private final Properties properties = new Properties();
+
+    {
+		try {
+			properties.load(mod_ExampleMod.class.getResourceAsStream("/exampleMod.properties"));
+		} catch (Throwable ignored) {
+		}
+	}
+
+    @Override
+    public String getVersion() {
+        return properties.getProperty("version");
+    }
+
+	@Override
+	public String getName() {
+		return properties.getProperty("name");
+	}
+
+	@Override
+    public void load() {
+        ExampleMod.LOGGER.info("Hello ModLoader world!");
+    }
+}
