+++ testing/minecraft/1.2.5/src/modloader/resources/exampleMod.properties
@@ -1,0 +1,2 @@
+name=Example Mod
+version=${version}
