+++ testing/minecraft/1.2.5/src/modloader/resources/exampleMod.transform
@@ -1,0 +1,1 @@
+com.example.example.mod.transform.GuiMainMenuTransform
